Created a new directory to put Image Detail Steps into, and broke them out into different (step) files.
