+++ ConnectSdk.OnlineTests/ImageDetails/When.cs
@@ -1,0 +1,18 @@
+﻿using TechTalk.SpecFlow;
+
+namespace GettyImages.Connect.Tests.ImageDetails
+{
+    [Binding]
+    [Scope(Feature = "Image Details")]
+    public class When
+    {
+        [When(@"I retrieve image details")]
+        [When(@"I retrieve details for the image")]
+        [When(@"I retrieve details for the images")]
+        public void WhenIRetrieveImageDetails()
+        {
+            ScenarioContext.Current.Add("task", ScenarioContext.Current.Get<Images>("request").ExecuteAsync());
+        }
+
+    }
+}
